tests/wpt/tests/css/css-overflow/scrollable-overflow-with-nested-elements-002.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Overflow: Scrollable Overflow Transform Rotate Nested Element</title>
<link rel="author" title="Shubham Gupta" href="mailto:shubham13297@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-overflow-3/#scrollable" />
<meta name="assert" content="Checks the scrollable overflow of parent with overflow auto and some nested element">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<style>
    .container {
        width: 100px;
        height: 100px;
        background: silver;
        scrollbar-width: none;
    }

    .list .item {
        width: 10px;
        height: 10px;
        background: lime;
        position: absolute;
        margin: 0%;
    }

    .list {
        position: relative;
        background: grey;
        scrollbar-width: none;
    }
</style>

<body onload="checkLayout('.container')">
    <div class="container" style="overflow-x: hidden;" data-expected-scroll-height="1010" data-expected-height="100" data-expected-width="100">
        <div class="list" data-expected-scroll-height="1010"
            data-expected-bounding-client-rect-height="0" data-expected-bounding-client-rect-width="100"
            data-expected-height="0" data-expected-width="100">
            <p class="item" style="top:0px;"></p>
            <p class="item" style="top:1000px;"></p>
        </div>
    </div>
    <div class="container" data-expected-scroll-height="1010" data-expected-height="100" data-expected-width="100">
        <div class="list" data-expected-scroll-height="1010"
            data-expected-bounding-client-rect-height="0" data-expected-bounding-client-rect-width="100"
            data-expected-height="0" data-expected-width="100">
            <p class="item" style="top:0px;"></p>
            <p class="item" style="top:1000px;"></p>
        </div>
    </div>
</body>
